requirements.txt
fastapi
uvicorn[standart]
sqlalchemy
python-multipart
pyjwt
asyncpg
passlib[bcrypt]
redis
celery
bcrypt==4.0.1
httpx
jinja2
alembic
dotenv